Optimised the code

+++ volume/volume.c
@@ -20,10 +20,9 @@
 
 // REQUIRED SUBHEADERS
 #include "../cvect.h"
-#include "../section/section.h"
 
 //FUNCTION TO FIND THE VOLUME OF PARALLELOPIPED
-float parellopiped(int *x, int *y, int *z)
+int parellopiped(int *x, int *y, int *z)
 {
-    return (float)abs(s3(x,y,z));
+    return abs(s3(x,y,z));
 }
